# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "后台管理系统 API",
        "title": "Admin Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/folders": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "创建文件夹",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin-server_internal_model.CreateFolderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FolderRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
        "/api/folders/tree": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "文件夹树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FolderTreeRespWrapper"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
        "/api/folders/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "更新文件夹",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件夹ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin-server_internal_model.UpdateFolderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FolderRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "删除文件夹",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件夹ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.MsgRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin-server_internal_model.CreateFolderReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "parentId": {
                    "type": "integer"
                }
            }
        },
...
        "admin-server_internal_model.FileFolder": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "admin-server_internal_model.FolderTreeNode": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin-server_internal_model.FolderTreeNode"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                }
            }
        },
...
        "admin-server_internal_model.UpdateFolderReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handler.ErrorRespWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "A0400"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.FolderRespWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "00000"
                },
                "data": {
                    "$ref": "#/definitions/admin-server_internal_model.FileFolder"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "internal_handler.FolderTreeRespWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "00000"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin-server_internal_model.FolderTreeNode"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "internal_handler.MsgRespWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "00000"
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "操作成功"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "格式: Bearer {accessToken}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}